+++ tests/__tests__/tests.test.ts
@@ -1,0 +1,6 @@
+import { sitka, store} from "../sitka-test"
+
+test(`get default null value from color module`, () => {
+  const colorState = store.getState()
+  console.log(colorState)
+})
